Extract logic to a new file to avoid circular dependencies
